Checking previously installed instance first

+++ inst_debian_wg.sh
@@ -6,12 +6,14 @@ SERVERPORT=51194
 SERVERPORTALT=53,51820
 FWDPORT=20000
 # -----
-systemctl stop wg-quick@wg0.service
-systemctl disable wg-quick@wg0.service
-/etc/wireguard/wg0_rules.sh -d
-rm /etc/sysctl.d/wg0.conf
-rm /etc/wireguard/wg0*
-sysctl --system
+if [ -f /etc/wireguard/wg0_rules.sh ]; then
+  systemctl stop wg-quick@wg0.service
+  systemctl disable wg-quick@wg0.service
+  /etc/wireguard/wg0_rules.sh -d
+  rm /etc/sysctl.d/wg0.conf
+  rm /etc/wireguard/wg0*
+  sysctl --system
+fi
 # -----
 apt-get install -y \
   wireguard wireguard-tools \
